# policy-engine/schemas/runtime_api_v1.openapi.json
{
  "components": {
    "schemas": {
      "ApiMeta": {
        "additionalProperties": false,
        "properties": {
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "source_kinds": {
            "items": {
              "const": "core_run",
              "type": "string"
            },
            "title": "Source Kinds",
            "type": "array"
          }
        },
        "required": [
          "request_id"
        ],
        "title": "ApiMeta",
        "type": "object"
      },
      "ArtifactContentPreview": {
        "additionalProperties": false,
        "properties": {
          "artifact_id": {
            "title": "Artifact Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "max_bytes": {
            "minimum": 1.0,
            "title": "Max Bytes",
            "type": "integer"
          },
          "media_type": {
            "title": "Media Type",
            "type": "string"
          },
          "mode": {
            "enum": [
              "json",
              "text",
              "binary"
            ],
            "title": "Mode",
            "type": "string"
          },
          "preview": {
            "title": "Preview"
          },
          "size_bytes": {
            "minimum": 0.0,
            "title": "Size Bytes",
            "type": "integer"
          },
          "truncated": {
            "default": false,
            "title": "Truncated",
            "type": "boolean"
          }
        },
        "required": [
          "artifact_id",
          "kind",
          "media_type",
          "mode",
          "size_bytes",
          "max_bytes"
        ],
        "title": "ArtifactContentPreview",
        "type": "object"
      },
      "ArtifactContentResponse": {
        "additionalProperties": false,
        "properties": {
          "artifact": {
            "$ref": "#/components/schemas/ArtifactContentPreview"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "required": [
          "meta",
          "artifact"
        ],
        "title": "ArtifactContentResponse",
        "type": "object"
      },
...
      "ArtifactLineageEdge": {
        "additionalProperties": false,
        "properties": {
          "child_artifact_id": {
            "title": "Child Artifact Id",
            "type": "string"
          },
          "parent_artifact_id": {
            "title": "Parent Artifact Id",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "parent_artifact_id",
          "child_artifact_id",
          "role"
        ],
        "title": "ArtifactLineageEdge",
        "type": "object"
      },
...
      "ArtifactLineageResponse": {
        "additionalProperties": false,
        "properties": {
          "lineage": {
            "$ref": "#/components/schemas/ArtifactLineageView"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "required": [
          "meta",
          "lineage"
        ],
        "title": "ArtifactLineageResponse",
        "type": "object"
      },
      "ArtifactLineageView": {
        "additionalProperties": false,
        "properties": {
          "corrupted_artifact_ids": {
            "items": {
              "type": "string"
            },
            "title": "Corrupted Artifact Ids",
            "type": "array"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/ArtifactLineageEdge"
            },
            "title": "Edges",
            "type": "array"
          },
          "is_complete": {
            "default": false,
            "title": "Is Complete",
            "type": "boolean"
          },
          "missing_artifact_ids": {
            "items": {
              "type": "string"
            },
            "title": "Missing Artifact Ids",
            "type": "array"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/ArtifactLineageNode"
            },
            "title": "Nodes",
            "type": "array"
          },
          "root_artifact_ids": {
            "items": {
              "type": "string"
            },
            "title": "Root Artifact Ids",
            "type": "array"
          },
          "total_edges": {
            "default": 0,
            "minimum": 0.0,
            "title": "Total Edges",
            "type": "integer"
          },
          "total_nodes": {
            "default": 0,
            "minimum": 0.0,
            "title": "Total Nodes",
            "type": "integer"
          },
          "total_size_bytes": {
            "default": 0,
            "minimum": 0.0,
            "title": "Total Size Bytes",
            "type": "integer"
          }
        },
        "title": "ArtifactLineageView",
        "type": "object"
      },
      "ArtifactManifestResponse": {
        "additionalProperties": false,
        "properties": {
          "artifact": {
            "$ref": "#/components/schemas/ArtifactManifestView"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "required": [
          "meta",
          "artifact"
        ],
        "title": "ArtifactManifestResponse",
        "type": "object"
      },
      "ArtifactManifestView": {
        "additionalProperties": false,
        "properties": {
          "artifact_id": {
            "title": "Artifact Id",
            "type": "string"
          },
          "byte_size": {
            "minimum": 0.0,
            "title": "Byte Size",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "inputs": {
            "items": {
              "$ref": "#/components/schemas/InputRef"
            },
            "title": "Inputs",
            "type": "array"
          },
          "integrity_sha256": {
            "title": "Integrity Sha256",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "media_type": {
            "title": "Media Type",
            "type": "string"
          },
          "producer_component": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Producer Component"
          },
          "producer_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Producer Version"
          },
          "schema_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Name"
          },
          "schema_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Version"
          }
        },
        "required": [
          "artifact_id",
          "kind",
          "media_type",
          "byte_size",
          "created_at",
          "integrity_sha256"
        ],
        "title": "ArtifactManifestView",
        "type": "object"
      },
      "ArtifactRef": {
        "additionalProperties": false,
        "properties": {
          "artifact_id": {
            "$ref": "#/components/schemas/ArtifactID"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "media_type": {
            "title": "Media Type",
            "type": "string"
          }
        },
        "required": [
          "artifact_id",
          "kind",
          "media_type"
        ],
        "title": "ArtifactRef",
        "type": "object"
      },
      "ArtifactSchemaResponse": {
        "additionalProperties": false,
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          },
          "schema": {
            "$ref": "#/components/schemas/ArtifactSchemaView"
          }
        },
        "required": [
          "meta",
          "schema"
        ],
        "title": "ArtifactSchemaResponse",
        "type": "object"
      },
      "ArtifactSchemaView": {
        "additionalProperties": false,
        "properties": {
          "artifact_id": {
            "title": "Artifact Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "media_type": {
            "title": "Media Type",
            "type": "string"
          },
          "schema_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Name"
          },
          "schema_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Version"
          },
          "top_level_keys": {
            "items": {
              "type": "string"
            },
            "title": "Top Level Keys",
            "type": "array"
          }
        },
        "required": [
          "artifact_id",
          "kind",
          "media_type"
        ],
        "title": "ArtifactSchemaView",
        "type": "object"
      },
...
      "GovernanceDebugResponse": {
        "additionalProperties": false,
        "properties": {
          "debug": {
            "$ref": "#/components/schemas/GovernanceDebugView"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "required": [
          "meta",
          "debug"
        ],
        "title": "GovernanceDebugResponse",
        "type": "object"
      },
      "GovernanceDebugView": {
        "additionalProperties": false,
        "properties": {
          "fallback_from_decision_packet": {
            "default": false,
            "title": "Fallback From Decision Packet",
            "type": "boolean"
          },
          "issues": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Issues",
            "type": "array"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "report_ref": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ArtifactRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "source_kind": {
            "const": "core_run",
            "title": "Source Kind",
            "type": "string"
          },
          "validation_trace": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation Trace"
          },
          "verdict": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verdict"
          }
        },
        "required": [
          "run_id",
          "source_kind"
        ],
        "title": "GovernanceDebugView",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InputRef": {
        "additionalProperties": false,
        "properties": {
          "artifact_id": {
            "$ref": "#/components/schemas/ArtifactID"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "artifact_id",
          "role"
        ],
        "title": "InputRef",
        "type": "object"
      },
      "NodeDebugResponse": {
        "additionalProperties": false,
        "properties": {
          "debug": {
            "$ref": "#/components/schemas/NodeDebugView"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "required": [
          "meta",
          "debug"
        ],
        "title": "NodeDebugResponse",
        "type": "object"
      },
      "NodeDebugView": {
        "additionalProperties": false,
        "properties": {
          "alias": {
            "title": "Alias",
            "type": "string"
          },
          "cache_bypasses": {
            "default": 0,
            "minimum": 0.0,
            "title": "Cache Bypasses",
            "type": "integer"
          },
          "cache_hits": {
            "default": 0,
            "minimum": 0.0,
            "title": "Cache Hits",
            "type": "integer"
          },
          "cache_stores": {
            "default": 0,
            "minimum": 0.0,
            "title": "Cache Stores",
            "type": "integer"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "record": {
            "$ref": "#/components/schemas/RunNodeRecord"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "source_kind": {
            "const": "core_run",
            "title": "Source Kind",
            "type": "string"
          },
          "timeline_events": {
            "items": {
              "$ref": "#/components/schemas/RunTimelineEvent"
            },
            "title": "Timeline Events",
            "type": "array"
          }
        },
        "required": [
          "run_id",
          "source_kind",
          "alias",
          "record"
        ],
        "title": "NodeDebugView",
        "type": "object"
      },
...
      "RunErrorView": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "node_alias": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Node Alias"
          },
          "source": {
            "enum": [
              "manifest",
              "workflow_report",
              "trace",
              "runtime"
            ],
            "title": "Source",
            "type": "string"
          },
          "timestamp": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          }
        },
        "required": [
          "source",
          "code",
          "message"
        ],
        "title": "RunErrorView",
        "type": "object"
      },
      "RunErrorsResponse": {
        "additionalProperties": false,
        "properties": {
          "errors": {
            "items": {
              "$ref": "#/components/schemas/RunErrorView"
            },
            "title": "Errors",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "meta",
          "run_id"
        ],
        "title": "RunErrorsResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Runtime API v1: run explorer, debug, and artifact inspector surfaces for PolicyOS.",
    "title": "PolicyOS Runtime API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/artifacts/{artifact_id}": {
      "get": {
        "operationId": "get_artifact_manifest",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactManifestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Artifact Manifest",
        "tags": [
          "runtime-artifacts"
        ]
      }
    },
    "/api/v1/artifacts/{artifact_id}/content": {
      "get": {
        "operationId": "get_artifact_content",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "title": "Artifact Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "max_bytes",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 2000000,
                  "minimum": 1024,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Bytes"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactContentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Artifact Content",
        "tags": [
          "runtime-artifacts"
        ]
      }
    },
    "/api/v1/artifacts/{artifact_id}/lineage": {
      "get": {
        "operationId": "get_artifact_lineage",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "title": "Artifact Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "max_depth",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 256,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Depth"
            }
          },
          {
            "in": "query",
            "name": "max_nodes",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 20000,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Nodes"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactLineageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Artifact Lineage",
        "tags": [
          "runtime-artifacts"
        ]
      }
    },
    "/api/v1/artifacts/{artifact_id}/schema": {
      "get": {
        "operationId": "get_artifact_schema",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactSchemaResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Artifact Schema",
        "tags": [
          "runtime-artifacts"
        ]
      }
    },
    "/api/v1/debug/runs/{run_id}/errors": {
      "get": {
        "operationId": "get_run_errors",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunErrorsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run Errors",
        "tags": [
          "runtime-debug"
        ]
      }
    },
    "/api/v1/debug/runs/{run_id}/governance": {
      "get": {
        "operationId": "get_governance_debug",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernanceDebugResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Governance Debug",
        "tags": [
          "runtime-debug"
        ]
      }
    },
    "/api/v1/debug/runs/{run_id}/nodes/{alias}": {
      "get": {
        "operationId": "get_node_debug",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "alias",
            "required": true,
            "schema": {
              "title": "Alias",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeDebugResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Node Debug",
        "tags": [
          "runtime-debug"
        ]
      }
    },
...
  }
}